+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.5.10 - 2022-
+- Added "who" CLI command to look up citation associated to named asteroid
+- Removed CLASS_TO_COMPLEX dictionary: complex is given in SsODNet
+
 # 1.5.9 - 2022-07-11
 - Fix highlighting of best-estimate parameters in datacloud queries
 - Fix bug in datacloud catalogue update